Spring-84
Algorithm: Division plus 1 this and write a bit less or more and multiply by this number back. If prime number -1 and divide. Repeat this many times. The number must be two n like this 2 ** n. E.g.: 32 bits 2 ** 34 and plus the same bits of long of the times: times (2**32)-1. 26-65535 bits. Deep: X + 25 bits. E.g.

Deep: 7

7+25=32 bits.

This program is for Quantum Supercomputers.